Image

+++ db/migrate/20190306060445_create_product.rb
@@ -2,7 +2,8 @@ class CreateProduct < ActiveRecord::Migration[5.2]
   def change
     create_table :products do |t|
     	t.string :name
-    	t.integer :price
+    	t.decimal :price
+
     	t.integer :quantity
     	t.string :description
     	t.integer :size
